Split ConversationPost into 3 seperate components
